更新 用户数据统计表 单项操作部分
